animal room - step 2

+++ addons/study/models/animal.py
@@ -6,4 +6,4 @@ class Animal(models.Model):
     _description = 'study_animal'
 
     name = fields.Char(required=True, string="Name")
-    total_number = fields.Float(string="Total Number")
+    total_number = fields.Integer(string="Total Number")
